+++ trip/src/main/java/dhbw/mosbach/trip/dtos/Route.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class Route {
-    private long startLat;
-    private long startLng;
-    private long endLat;
-    private long endLng;
+    private double startLat;
+    private double startLng;
+    private double endLat;
+    private double endLng;
 }
